// src/types-api.d.ts
export type Root = Root2[]

export interface Root2 {
  flags: Flags
  name: Name
  capital: string[]
  subregion: string
}

export interface Flags {
  png: string
  svg: string
  alt: string
}

export interface Name {
  common: string
  official: string
  nativeName: NativeName
}

export interface NativeName {
  ell?: Ell
  tur?: Tur
  ara?: Ara
  eng?: Eng
  tir?: Tir
  ita?: Ita
  lat?: Lat
  rar?: Rar
  som?: Som
  spa?: Spa
  rus?: Rus
  tuk?: Tuk
  sqi?: Sqi
  hrv?: Hrv
  por?: Por
  tet?: Tet
  fra?: Fra
  kon?: Kon
  lin?: Lin
  aze?: Aze
  swa?: Swa
  smo?: Smo
  vie?: Vie
  prs?: Prs
  pus?: Pus
  lav?: Lav
  kir?: Kir
  pol?: Pol
  cha?: Cha
  lit?: Lit
  hye?: Hye
  nrf?: Nrf
  tgk?: Tgk
  amh?: Amh
  ber?: Ber
  mey?: Mey
  mri?: Mri
  nzs?: Nzs
  msa?: Msa
  niu?: Niu
  ron?: Ron
  nor?: Nor
  bel?: Bel
  ces?: Ces
  slk?: Slk
  urd?: Urd
  aym?: Aym
  que?: Que
  kal?: Kal
  hun?: Hun
  zdj?: Zdj
  ben?: Ben
  tkl?: Tkl
  fij?: Fij
  hif?: Hif
  zho?: Zho
  div?: Div
  gle?: Gle
  est?: Est
  grn?: Grn
  afr?: Afr
  nbl?: Nbl
  nso?: Nso
  sot?: Sot
  ssw?: Ssw
  tsn?: Tsn
  tso?: Tso
  ven?: Ven
  xho?: Xho
  zul?: Zul
  bis?: Bis
  fin?: Fin
  swe?: Swe
  pih?: Pih
  bjz?: Bjz
  mon?: Mon
  tvl?: Tvl
  nep?: Nep
  bul?: Bul
  glv?: Glv
  dzo?: Dzo
  khm?: Khm
  hat?: Hat
  kat?: Kat
  run?: Run
  mfe?: Mfe
  nya?: Nya
  hmo?: Hmo
  tpi?: Tpi
  deu?: Deu
  heb?: Heb
  tam?: Tam
  ton?: Ton
  nld?: Nld
  uzb?: Uzb
  mlg?: Mlg
  mah?: Mah
  crs?: Crs
  jpn?: Jpn
  cat?: Cat
  fas?: Fas
  mlt?: Mlt
  kor?: Kor
  jam?: Jam
  dan?: Dan
  fil?: Fil
  bar?: Bar
  pov?: Pov
  her?: Her
  hgm?: Hgm
  kwn?: Kwn
  loz?: Loz
  ndo?: Ndo
  bos?: Bos
  srp?: Srp
  cnr?: Cnr
  cal?: Cal
  ukr?: Ukr
  arc?: Arc
  ckb?: Ckb
  fao?: Fao
  gil?: Gil
  kaz?: Kaz
  isl?: Isl
  pau?: Pau
  gsw?: Gsw
  roh?: Roh
  mya?: Mya
  tha?: Tha
  pap?: Pap
  lao?: Lao
  hin?: Hin
  ltz?: Ltz
  sag?: Sag
  nfr?: Nfr
  nno?: Nno
  nob?: Nob
  smi?: Smi
  bwg?: Bwg
  kck?: Kck
  khi?: Khi
  ndc?: Ndc
  nde?: Nde
  sna?: Sna
  toi?: Toi
  zib?: Zib
  slv?: Slv
  lua?: Lua
  ind?: Ind
  nau?: Nau
  kin?: Kin
  mkd?: Mkd
  sin?: Sin
}